feat(TrackSegment): Pass 2 — typed placement data pipeline
Migrate usePlacementData and populateSide/ZSteps modules to TypeScript with
PopulateSideArgs / PlacementData interfaces. Keeps decoration bag keys always
defined to prevent the historical undefined-identifier crashes.
